+++ fbgemm_gpu/CMakeLists.txt
@@ -289,7 +289,7 @@ set(fbgemm_gpu_sources_cpu
     codegen/embedding_bounds_check_host_cpu.cpp
     backend/cpu/native/cpu_utils.cpp
     backend/cpu/native/jagged_tensor_ops_cpu.cpp
-    src/input_combine_cpu.cpp
+    backend/cpu/native/input_combine_cpu.cpp
     backend/cpu/native/layout_transform_ops_cpu.cpp
     backend/cpu/native/quantize_ops_cpu.cpp
     backend/cpu/native/sparse_ops_cpu.cpp)
@@ -301,19 +301,20 @@ if(NOT FBGEMM_CPU_ONLY)
     codegen/embedding_forward_quantized_host.cpp
     codegen/embedding_backward_dense_host.cpp
     codegen/embedding_bounds_check_host.cpp
-    src/cumem_utils_host.cpp
+    backend/cuda/native/cumem_utils.host.cpp
     backend/cuda/native/layout_transform_ops_gpu.cpp
-    src/permute_pooled_embedding_ops_gpu.cpp
-    src/permute_pooled_embedding_ops_split_gpu.cpp
+    backend/cpu/native/permute_pooled_embedding_ops_gpu.cpp
+    backend/cuda/native/permute_pooled_embedding_ops_split_gpu.cpp
     src/permute_pooled_embedding_ops_split_cpu.cpp
     backend/cuda/native/quantize_ops_gpu.cpp
     backend/cuda/native/sparse_ops_gpu.cpp
-    src/split_table_batched_embeddings.cpp
-    src/metric_ops_host.cpp)
+    backend/cuda/native/split_table_batched_embeddings.cpp
+    backend/cuda/native/metric_ops.host.cpp)
 
   if(NVML_LIB_PATH)
-    list(APPEND fbgemm_gpu_sources_cpu src/merge_pooled_embeddings_cpu.cpp
-         src/merge_pooled_embeddings_gpu.cpp)
+    list(APPEND fbgemm_gpu_sources_cpu
+         backend/cpu/native/merge_pooled_embeddings_cpu.cpp
+         backend/cuda/native/merge_pooled_embeddings_gpu.cpp)
   endif()
 endif()
 
@@ -324,17 +325,17 @@ set_source_files_properties(
 if(NOT FBGEMM_CPU_ONLY)
   set(fbgemm_gpu_sources_gpu
       codegen/embedding_bounds_check.cu
-      src/cumem_utils.cu
+      backend/cuda/native/cumem_utils.cu
       backend/cuda/native/histogram_binning_calibration_ops.cu
       backend/cuda/native/jagged_tensor_ops.cu
       backend/cuda/native/layout_transform_ops.cu
-      src/permute_pooled_embedding_ops.cu
-      src/permute_pooled_embedding_ops_split.cu
+      backend/cuda/native/permute_pooled_embedding_ops.cu
+      backend/cuda/native/permute_pooled_embedding_ops_split.cu
       backend/cuda/native/quantize_ops.cu
       backend/cuda/native/sparse_ops.cu
-      src/split_embeddings_cache_cuda.cu
+      backend/cuda/native/split_embeddings_cache_cuda.cu
       backend/cuda/native/split_embeddings_utils.cu
-      src/metric_ops.cu)
+      backend/cuda/native/metric_ops.cu)
 
   set_source_files_properties(
     ${fbgemm_gpu_sources_gpu} PROPERTIES COMPILE_OPTIONS
