+++ fbgemm_gpu/docs/Doxyfile.in
@@ -916,10 +916,10 @@ INPUT                  = "../include/fbgemm_gpu" \
                          "../include/fbgemm_gpu/sparse_ops.h" \
                          "../fbgemm_gpu/backend/cuda/native/quantize_ops.cu" \
                          "../backend/cpu/native/quantize_ops_cpu.cpp" \
-                         "../src/split_table_batched_embeddings.cpp" \
+                         "../backend/cuda/native/split_table_batched_embeddings.cpp" \
                          "../backend/cuda/native/jagged_tensor_ops.cu" \
                          "../backend/cpu/native/jagged_tensor_ops_cpu.cpp" \
-                         "../src/cumem_utils.h" \
+                         "../include/fbgemm_gpu/cumem_utils.h" \
                          "../include/fbgemm_gpu/input_combine.h" \
                          "../backend/cuda/native/layout_transform_ops.cu" \
                          "../backend/cpu/native/layout_transform_ops_cpu.cpp" \
@@ -931,7 +931,7 @@ INPUT                  = "../include/fbgemm_gpu" \
                          "../codegen/embedding_bounds_check_host_cpu.cpp" \
                          "../codegen/embedding_backward_dense_host.cpp" \
                          "../codegen/embedding_forward_template_helpers.cuh " \
-                         "../src/permute_pooled_embedding_ops_gpu.cpp" \
+                         "../backend/cpu/native/permute_pooled_embedding_ops_gpu.cpp" \
 
 # This tag can be used to specify the character encoding of the source files
 # that doxygen parses. Internally doxygen uses the UTF-8 encoding. Doxygen uses
